changed all variables to private double

+++ src/Employment2016.java
@@ -1,38 +1,38 @@
 public class Employment2016 {
 
-    private int totalLaborForce;
-    private int employedLaborForce;
-    private int unemployedLaborForce;
+    private double totalLaborForce;
+    private double employedLaborForce;
+    private double unemployedLaborForce;
     private double unemployedPercent;
 
-    public Employment2016(int totalLaborForce, int employedLaborForce, int unemployedLaborForce, double unemployedPercent) {
+    public Employment2016(double totalLaborForce, double employedLaborForce, double unemployedLaborForce, double unemployedPercent) {
         this.totalLaborForce = totalLaborForce;
         this.employedLaborForce = employedLaborForce;
         this.unemployedLaborForce = unemployedLaborForce;
         this.unemployedPercent = unemployedPercent;
     }
 
-    public int getTotalLaborForce() {
+    public double getTotalLaborForce() {
         return totalLaborForce;
     }
 
-    public void setTotalLaborForce(int totalLaborForce) {
+    public void setTotalLaborForce(double totalLaborForce) {
         this.totalLaborForce = totalLaborForce;
     }
 
-    public int getEmployedLaborForce() {
+    public double getEmployedLaborForce() {
         return employedLaborForce;
     }
 
-    public void setEmployedLaborForce(int employedLaborForce) {
+    public void setEmployedLaborForce(double employedLaborForce) {
         this.employedLaborForce = employedLaborForce;
     }
 
-    public int getUnemployedLaborForce() {
+    public double getUnemployedLaborForce() {
         return unemployedLaborForce;
     }
 
-    public void setUnemployedLaborForce(int unemployedLaborForce) {
+    public void setUnemployedLaborForce(double unemployedLaborForce) {
         this.unemployedLaborForce = unemployedLaborForce;
     }
